+++ docs/ui/components/TableOfContents/TableOfContents.test.tsx
@@ -1,5 +1,5 @@
 import { jest } from '@jest/globals';
-import { act, fireEvent } from '@testing-library/react';
+import { act, fireEvent, screen } from '@testing-library/react';
 import GithubSlugger from 'github-slugger';
 import { createRef, type RefObject } from 'react';
 
@@ -92,7 +92,7 @@ describe('TableOfContents', () => {
     const { contentRef } = createContentRef();
     const tocRef = createRef<TableOfContentsHandles>();
 
-    const { getByText } = renderWithHeadings(
+    renderWithHeadings(
       <HeadingsContext.Provider value={headingManager}>
         <TableOfContents ref={tocRef} headingManager={headingManager} contentRef={contentRef} />
       </HeadingsContext.Provider>
@@ -100,16 +100,13 @@ describe('TableOfContents', () => {
 
     act(() => tocRef.current?.handleContentScroll?.(0));
     expect(tocRef.current).not.toBeNull();
-    const introText = getByText('Intro');
-    expect(introText).toHaveClass('text-link!');
+    expect(screen.getByText('Intro')).toHaveClass('text-link!');
 
     act(() => tocRef.current?.handleContentScroll?.(950));
-    const middleText = getByText('Middle');
-    expect(middleText).toHaveClass('text-link!');
+    expect(screen.getByText('Middle')).toHaveClass('text-link!');
 
     act(() => tocRef.current?.handleContentScroll?.(1500));
-    const endText = getByText('End');
-    expect(endText).toHaveClass('text-link!');
+    expect(screen.getByText('End')).toHaveClass('text-link!');
   });
 
   test('scrolls to align heading with activation line on click', () => {
@@ -124,15 +121,13 @@ describe('TableOfContents', () => {
     const { contentRef, scrollTo, scrollRef } = createContentRef(800, 1600);
     const tocRef = createRef<TableOfContentsHandles>();
 
-    const { getByText } = renderWithHeadings(
+    renderWithHeadings(
       <HeadingsContext.Provider value={headingManager}>
         <TableOfContents ref={tocRef} headingManager={headingManager} contentRef={contentRef} />
       </HeadingsContext.Provider>
     );
 
-    act(() => {
-      fireEvent.click(getByText('Middle'));
-    });
+    fireEvent.click(screen.getByText('Middle'));
 
     expect(scrollTo).toHaveBeenCalledWith(
       expect.objectContaining({