+++ packages/@expo/cli/src/run/android/resolveGradlePropsAsync.ts
@@ -53,7 +53,7 @@ export async function resolveGradlePropsAsync(
 
   let apkVariantDirectory: string;
   if (parts.length > 0) {
-    const flavorPath = parts[0].toLowerCase() + parts.slice(1).join('');
+    const flavorPath = parts[0]!.toLowerCase() + parts.slice(1).join('');
     apkVariantDirectory = path.join(apkDirectory, flavorPath, buildType);
   } else {
     apkVariantDirectory = path.join(apkDirectory, buildType);