+++ packages/@expo/config-plugins/src/ios/codeMod.ts
@@ -194,7 +194,7 @@ export function findSwiftFunctionCodeBlock(contents: string, selector: string): 
       continue;
     }
     for (let i = 0; i < argLabels.length; ++i) {
-      if (argLabels[i] !== params[i].argumentLabel) {
+      if (argLabels[i] !== params[i]?.argumentLabel) {
         continue;
       }
     }
@@ -215,7 +215,7 @@ export function findSwiftFunctionCodeBlock(contents: string, selector: string): 
 
 function parseSwiftFunctionParam(paramTuple: string): SwiftFunctionParam {
   const semiIndex = paramTuple.indexOf(':');
-  const [argumentLabel, parameterName] = paramTuple.substring(0, semiIndex).split(/\s+/);
+  const [argumentLabel = '', parameterName = ''] = paramTuple.substring(0, semiIndex).split(/\s+/);
   const typeString = paramTuple.substring(semiIndex + 1).trim();
   return {
     argumentLabel,