+++ packages/@expo/prebuild-config/src/plugins/unversioned/expo-splash-screen/InterfaceBuilder.ts
@@ -307,26 +307,33 @@ export function removeImageFromSplashScreen(
   xml: IBSplashScreenDocument,
   { imageName }: { imageName: string }
 ) {
-  const mainView = xml.document.scenes[0].scene[0].objects[0].viewController[0].view[0];
+  const mainView = xml.document.scenes[0]?.scene[0]?.objects[0]?.viewController[0]?.view[0];
 
   debug(`Remove all splash screen image elements`);
 
-  removeExisting(mainView.subviews[0].imageView, IMAGE_ID);
+  if (mainView != null) {
+    if (mainView.subviews[0] != null) {
+      removeExisting(mainView.subviews[0].imageView, IMAGE_ID);
+    }
 
-  // Remove Constraints
-  getAbsoluteConstraints(IMAGE_ID, CONTAINER_ID).forEach((constraint) => {
-    // <constraint firstItem="EXPO-SplashScreen" firstAttribute="top" secondItem="EXPO-ContainerView" secondAttribute="top" id="2VS-Uz-0LU"/>
-    const constrainsArray = mainView.constraints[0].constraint;
-    removeExisting(constrainsArray, constraint);
-  });
+    // Remove Constraints
+    getAbsoluteConstraints(IMAGE_ID, CONTAINER_ID).forEach((constraint) => {
+      // <constraint firstItem="EXPO-SplashScreen" firstAttribute="top" secondItem="EXPO-ContainerView" secondAttribute="top" id="2VS-Uz-0LU"/>
+      if (mainView.constraints[0] != null) {
+        removeExisting(mainView.constraints[0].constraint, constraint);
+      }
+    });
+  }
 
   // Remove resource
-  xml.document.resources[0].image = xml.document.resources[0].image ?? [];
-  const imageSection = xml.document.resources[0].image;
+  if (xml.document.resources[0] != null) {
+    xml.document.resources[0].image = xml.document.resources[0].image ?? [];
+    const imageSection = xml.document.resources[0].image;
 
-  const existingImageIndex = imageSection.findIndex((image) => image.$.name === imageName);
-  if (existingImageIndex && existingImageIndex > -1) {
-    imageSection?.splice(existingImageIndex, 1);
+    const existingImageIndex = imageSection.findIndex((image) => image.$.name === imageName);
+    if (existingImageIndex && existingImageIndex > -1) {
+      imageSection?.splice(existingImageIndex, 1);
+    }
   }
   return xml;
 }
@@ -362,11 +369,12 @@ export function applyImageToSplashScreenXML(
     imageWidth?: number;
   }
 ): IBSplashScreenDocument {
-  const mainView = xml.document.scenes[0].scene[0].objects[0].viewController[0].view[0];
+  const mainView = xml.document.scenes[0]?.scene[0]?.objects[0]?.viewController[0]?.view[0];
+  const rect = mainView?.rect[0];
   const width = enableFullScreenImage ? 414 : imageWidth;
   const height = enableFullScreenImage ? 736 : imageWidth;
-  const x = enableFullScreenImage ? 0 : (mainView.rect[0].$.width - width) / 2;
-  const y = enableFullScreenImage ? 0 : (mainView.rect[0].$.height - height) / 2;
+  const x = enableFullScreenImage || rect == null ? 0 : (rect.$.width - width) / 2;
+  const y = enableFullScreenImage || rect == null ? 0 : (rect.$.height - height) / 2;
 
   const imageView: IBImageView = {
     $: {
@@ -391,68 +399,77 @@ export function applyImageToSplashScreenXML(
     ],
   };
 
-  // Add ImageView
-  ensureUniquePush(mainView.subviews[0].imageView, imageView);
-
-  mainView.constraints[0].constraint = [];
-
-  // Add Constraints
-  getAbsoluteConstraints(IMAGE_ID, CONTAINER_ID, enableFullScreenImage).forEach(
-    (constraint: IBConstraint) => {
-      const constrainsArray = mainView.constraints[0].constraint;
-      ensureUniquePush(constrainsArray, constraint);
+  if (mainView != null) {
+    // Add ImageView
+    if (mainView.subviews[0] != null) {
+      ensureUniquePush(mainView.subviews[0].imageView, imageView);
     }
-  );
 
-  // Clear existing images
-  xml.document.resources[0].image = [];
+    if (mainView.constraints[0] != null) {
+      mainView.constraints[0].constraint = [];
+    }
 
-  // Add resource
-  const imageSection = xml.document.resources[0].image;
-  imageSection.push({
-    $: {
-      name: imageName,
-      width,
-      height,
-    },
-  });
+    // Add Constraints
+    getAbsoluteConstraints(IMAGE_ID, CONTAINER_ID, enableFullScreenImage).forEach(
+      (constraint: IBConstraint) => {
+        if (mainView.constraints[0] != null) {
+          ensureUniquePush(mainView.constraints[0].constraint, constraint);
+        }
+      }
+    );
 
-  // Clear existing color
-  mainView.color = [];
-  // Add background color
-  const colorSection = mainView.color;
+    // Clear existing color
+    mainView.color = [];
+    // Add background color
+    const colorSection = mainView.color;
 
-  colorSection.push({
-    $: {
-      key: 'backgroundColor',
-      name: 'SplashScreenBackground',
-    },
-  });
-
-  // Clear existing named colors
-  xml.document.resources[0].namedColor = [];
-
-  const namedColorSection = xml.document.resources[0].namedColor;
-  // Add background named color reference
-  const color = parseColor(backgroundColor);
-
-  namedColorSection.push({
-    $: {
-      name: 'SplashScreenBackground',
-    },
-    color: [
-      {
-        $: {
-          alpha: '1.000',
-          blue: color.rgb.blue,
-          green: color.rgb.green,
-          red: color.rgb.red,
-          customColorSpace: 'sRGB',
-          colorSpace: 'custom',
-        },
+    colorSection.push({
+      $: {
+        key: 'backgroundColor',
+        name: 'SplashScreenBackground',
       },
-    ],
-  });
+    });
+  }
+
+  if (xml.document.resources[0] != null) {
+    // Clear existing images
+    xml.document.resources[0].image = [];
+
+    // Add resource
+    const imageSection = xml.document.resources[0].image;
+    imageSection.push({
+      $: {
+        name: imageName,
+        width,
+        height,
+      },
+    });
+
+    // Clear existing named colors
+    xml.document.resources[0].namedColor = [];
+
+    const namedColorSection = xml.document.resources[0].namedColor;
+    // Add background named color reference
+    const color = parseColor(backgroundColor);
+
+    namedColorSection.push({
+      $: {
+        name: 'SplashScreenBackground',
+      },
+      color: [
+        {
+          $: {
+            alpha: '1.000',
+            blue: color.rgb.blue,
+            green: color.rgb.green,
+            red: color.rgb.red,
+            customColorSpace: 'sRGB',
+            colorSpace: 'custom',
+          },
+        },
+      ],
+    });
+  }
 
   return xml;
 }