+++ packages/create-expo/src/Template.ts
@@ -244,7 +244,9 @@ export async function getTemplateFilesToRenameAsync({
   let config = userConfig ?? defaultRenameConfig;
 
   // Strip comments, trim whitespace, and remove empty lines.
-  config = config.map((line) => line.split(/(?<!\\)#/, 2)[0].trim()).filter((line) => line !== '');
+  config = config
+    .map((line) => line.split(/(?<!\\)#/, 2)[0]?.trim() ?? '')
+    .filter((line) => line !== '');
 
   return await glob(config, {
     cwd,