+++ packages/expo-ui/src/jetpack-compose/index.ts
@@ -1,6 +1,8 @@
 import './MaterialSymbolsAssetsTransformer.fx';
 
 export * from './AlertDialog';
+export * from './Badge';
+export * from './BadgedBox';
 export { BasicAlertDialog, type BasicAlertDialogProps } from './BasicAlertDialog';
 export * from './Card';
 export * from './Checkbox';
@@ -33,6 +35,7 @@ export * from './PullToRefreshBox';
 export * from './RadioButton';
 export * from './Surface';
 export { type TextProps, Text } from './Text';
+export * from './Tooltip';
 
 export * from './AnimatedVisibility';
 export * from './Box';